(#508) Remove previous package attempt on uninstall
There are times where we have created a package result
before attempting the uninstall of a package, when we
come across scenarios for these we need to remove any
previously acquired package to prevent further failures
from occurring during the uninstall, or running of uninstall
scripts.

This was noticed when the user had requested to uninstall
multiple packages, where one package was specified and at
the same time was a dependency on a different package.
When this happens, the first attempt may fail due to it being
a dependency, but once the package referencing this package
is uninstalled while removing all dependencies, the uninstall would
still fail due to the previously cached result would have missing information
about the package.