(#508) Collect all specified package names before uninstall
This commit updates the handling of collecting information
about all package names the user has specified on the
command line before attempting to do any uninstalls, or
resolving any of its dependencies.

This allows us to make a better decision about which packages
need to be installed before others.

The code that was added here ensures that if the user specified
two packages, where one was a dependency on the other the package
having the dependency will be attempted to be uninstalled first.