(tests) Fix check for errors loading signed files
This test was just too fragile and not suitable for this use case by the
looks of things. Slight environmental differences would make it fail or
pass, so redesigning these ones to be a little more robust and only
error if the import / running of the script actually fails, not just if
they leak any errors that have already been handled into $error.