(#508) Add warning about filtering of packages
When running choco search, there are several options that would have
previously caused packages to be filtered on the server side, before
being returned to the client. The options were --approved-only,
--download-cache, and --not-broken.  The decision was taken to no
longer have these options passed to the server, since they can't
(currently) be respected on a V3 feed.  Instead, the filtering is
applied on the client side, within Chocolatey CLI.  The end result is
that it is possible to get incomplete pages of data returned, as some
packages may be filtered out on the client, and no additional packages
will be requested.

This commit adds a warning, to inform the user that this may be
happening.