(maint) Mark tests broken in test kitchen as foss only
There are a few E2E tests that we run through test kitchen
that are failing when running on a licensed extension, however
these have not been reproducible outside of test kitchen as of
this date.

As such we mark these tests as only available for Open Source where
they run as expected. Further investigations may be needed to figure
out the root cause of the issue.