(#508) Add warning result if before modify script fails
This commit updates the logic about running the Chocolatey
Before Modify script to include any messages outputted as part
of this run, to be in the result of the uninstall and/or the upgrade
results of the operation.

This allows a warning to be outputted when a before modify script
fails by any reason, which allows the user again to be notified of such
failures more easily as it is listed at the end of any operation.