(#508) Fix package installation order
The commit adds new removes the custom comparer to
instead handle the sorting of the packages that we need
to uninstall in its own method.

This is done as the normal comparer has limits that could
not be overridden and would not sort packages correctly
if there was too many packages installed at the same time.