(#3150) Update to latest Chocolatey-NuGet.Core
This new package version includes a fix to ensure that packages are
operated on in the correct order, specifcally when handling the
Chocolatey Products.  This will ensure for example that Chocolatey CLI
is upgraded before Chocolatey Agent is upgraded.  If this doesn't
happen, then there are known issues that can happen with restarting the
Chocolatey Agent Windows Service, since the new Chocolatey CLI
components haven't yet been placed on disk.

This commit updates the Chocolatey-NuGet.Core files on disk in the lib
folder, as well as updates the credits file to indicate a new version
is being used, and also updates the project references to new the new
assemblies.