(#508) Update snapshot service to respect version normalization
This commit updates the file snapshot service to honor version normalization
with fallbacks to old version methods.

This is done as we are moving towards normalizing version numbers in most of
the code, which causes a conflict when installing, upgrading or uninstalling
packages that sometimes it would use non-normalized version number,
and other times it would use normalized version numbers.
Now with this change it will prefer to use normalized version number, but will
continue to read/update package versions that was installed using
non-normalized version numbers.