(#508) Remove nupkg file on partial file removals
This commit updates the uninstall process to ensure that the nupkg
file is removed, even in the cases where only partial file removals has
happened. This will prevent the package from being listed as installed
when partial removal is done.

While the original code does not pose any problems with normal usage
of CLI, it will not remove the nuget packages when the package has been
modified, like what Licensed Extension does when optimizing installations.