(build) Ensure correct MSBuild ToolPath is set
While doing the 1.4.0 release of Chocolatey CLI, there was an error in
TeamCity when running the build.  This was due to the fact that we have
introduced SonarQube into the build process, and this only works with a
specific version of MSBuild.  In other places, Chocolatey.Cake.Recipe
was already specifically setting the ToolPath as done in this commit,
however, this was not done in the support branch.