(test) Restore usage of --reduce-nupkg-only option
During PR review, there was a misunderstanding as to the intention of
the test, partially due to a comment that had been left in the code,
and also mainly due to time constraints to get the PR merged.

This commit adds back in the usage of the reduce-nupkg-only option,
which will prevent Package Optimizer from cleaning up files during the
tests when running under the context of CLE, but will still result in
the nupkg itself being reduced.