(#508) Do case insensitive comparison on package ids
This commit updates the Equals comparisons in the nuget
service to be case insensitive.

This is done as without this change the user will always
have to pass in the identifier in the exact casing that
is returned from the nuget server, causing issues especially
during upgrade scenarios.

Additionally, Equals and Compare methods has been added
to the banned API list to prevent the use
of these without specifying a StringComparison.