(#158) Restore --source for choco list
Since alternate sources can list installed packages separately to our
normal package list, we need to allow list to specify --source.
This will have no effect for normal sources, but will select alternate
source runners if specified.

Additionally, perform early checks for each source runner as to which operations they support rather than rely on the runner itself to throw
exceptions when they don't support something. This was implemented in a
way that allows us to gut ISourceRunner in a later version and fall back
to having the supported methods for each source runner enforced by the
type system, so we have a clear indication as to which source runners
support each kind of operation via an appropriate interface cast.