(maint) Update Pester test for Pester v5 norms
Pester v5 introduced the BeforeDiscovery block, and any module that is
loaded in this context is not loaded when tests actually run. By default
anything outside of any other block run at this time, but to be
explicit, we should put them in a BeforeDiscovery block. This commit
removes the unneeded Import-Module statements from files we're editting
for the PR, as well as moves any variable assignments into
BeforeDiscovery so they're available and explicit.