(#158) Refactoring based on pairing session
During a pairing session between Kim and Gary, some additional changes
were made to further refactor the ISourceRunner. To make the
functionality work longer term, with intended upcoming changes, the
Container registrations for the ISourceRunners was changed, so that
mulitple casting to other interfaces would no longer be required.

In addition, methods that are no longer required on each SourceRunner
were removed, rather than leaving them to never get called again. This
is very much a breaking change, but it was felt that this was the right
time to make this change.