(#508) Ensure correct configuration object is used
When we first brought in the new NuGet.Client, a change was made to how
the callback for the HandlePackageResult was done.  Previously, this
didn't receive a ChocolateyConfiguration object from the callback, but
then it did.  As a result, the wrong instance of the
ChocolateyConfiguration object was sent along to continue the
operation, which meant all the environment information had not been set
correctly.  This caused a problem when trying to install the cygwin
package since there is a check against what operating system the
process is being run on, which hadn't been set correctly.

This came up when running choco install chere --source cygwin, which
failed to install Cygwin correctly.  This change has been in the code
base a while, and it was only when testing the installation of a
package from an alternative source that this came up.