(#3165) Update count method to support nuget service
This commit updates the Count method that is used to get
the amount of packages available on normal and alternative
sources to include support for the nuget service.

During a previous refactor we had removed the support of
acquiring packages on normal sources by a mistake, which
causes an issue with Chocolatey GUI to fail due to no available
source runner can be found when dealing with counting packages.
This was fixed by including an explicit call to the nuget service when
norrmal sources are being used.