(#948) Add user as sensitive argument
Since user is one half of the information that is needed for a
credential, let's treat the user as sensitive information, in the same
way that we treat password.

This commit adds the various forms that the user option can be passed
to the Chocolatey CLI, to ensure that it is always caught.