(#3183) Ensure ChocolateySourceCacheContext in use
In order to prevent a 1 second delay being imposed on each outgoing
query, we need to make sure that an instance of a
ChocolateySourceCacheContext is passed into every method that is called
within the Chocolatey.NuGet.Client libraries. In order for this to work
changes were needed to be made in the Chocolatey.NuGet.Client libraries
to provide the overrides for all of these methods.

This commit addresses issues that were reported with the 2.0.0 release
of Chocolatey CLI, where performance of commands was significantly
slower than on v1.4.0.