(#3125) Use ChocolateySourceCacheContext for search
In order to be consistent with how queries are cached when doing install
and upgrade commands, this commit switches to passing in an instance
of the ChocolateySourceCacheContext object, so that the choco search
command makes use of the same cached query results.  This will mean
that all commands return consistent results.  It does mean that we will
need to provide an ability to by-pass/purge these caches, but that will
be part of another scope of work.