(#3196) Pin add/remove use exact package ID
For the pin commands add and remove, find the package to pin/unpin
using an exact list, not a normal search, to ensure that only an exact
matching package ID is returned, and therefore pinned/unpinned.

A test is also added for this, using the mingw and gstreamer-mingw
package IDs, which is what was reported causing problems in the issue.