(#3182) Fix rule distinction comparison
When finding the distinct rules that are available in Chocolatey CLI
and its extensions, the equals and GetHashCode was incorrectly
implemented which in some cases causes the incorrect rules to be
returned, or not returned.

This commit updates these two methods to handle the check for
getting the hash code, and comparing immutable rules in the way
that we expect them to behave.