(#3190) Fix threshold warning missing on multi-source
There was a threshold warning added to Chocolatey CLI v2.0.0 that
would warn users about searches issued when one of the sources used
has equal or more packages available than what the threshold is set to.

Unfortunately this was only working when a single source was being used
in Chocolatey CLI v2.0.0.
The changes in this commit makes the necessary changes to have this
functionality work as expected again for multiple sources.