(maint) Reset paging configuration variables
When running integration tests the configuration items
that are used to determine paging was never reset as part
of getting the baseline configuration.

This caused multiple tests to fail when the new test that explicitly
tests using paging was added.

The changes here makes the necessary changes to the baseline
configuration to set the default values to prevent other tests to
be affected by this change.