(maint) Update to latest Chocolatey.NuGet.Client
Start using 3.3.0, which includes a fix for a bug that was introduced
with regard to long file paths when doing oData queries, and caching of
those results. In addition, a change was made to allow overridding of
how proxy configuration is done within choco.exe, to allow full support
for all proxy mechanisms.