(maint) Add structure for outputting help
This changes how the help output is implemented in
Chocolatey CLI.
It adds the handling of outputting the majority of the help
output in the class called ChocolateyCommandBase that can
be used to writing help documentation easier, and instead of
thinking about the structure of the output instead needs to
override specific methods to give the information that is needed.

Some defaults are additionally handled as well to make common
output be easier and more maintainable.