(maint) Remove env variable for release version
This was added in this commit:

https://github.com/chocolatey/choco/commit/e991fbc6c64a09b7ba8f3f0f346ba2724823da20

after the concept of package.Version.PackageReleaseVersion was added
onto the SemanticVersion class within the NuGet.Core assembly.  This
concept hasn't been brought over to the Chocolatey.Nuget.Client
libraries, and as such, the addition of this particular environment
variable is not required, since it is no different to the variables
that are already being emitted.  The suggestion going forward is likely
to be that package maintainers use the build metadata within SemVer v2
to indicate that a change has been made to the packaging files, rather
than the underlying application.  This concept won't see adoption until
SemVer v2 is added to CCR though.