(#3174) Store non-normalized package version
In a previous change we normalized all versions that was being used
in Chocolatey CLI. This caused certain packages that uses non-normalized
version numbers to start failing to look up additional information in other
products.

To work around this problem we need to store the non-normalized version
of the package when scripts are being ran. This is done by adding a new
environment variable that can be used, this environment variable is considered
private and is not for public consumption.