(#3224) Add Assembly Loaded configuration option
In order to determine if the Licensed Assembly has been loaded, we
either need to validate the license every time, or add it somewhere that
we can access more easily. This commit adds a `IsLicensedAssemblyLoaded`
boolean to the `InformationCommand` property of
`ChocolateyConfiguration` beside `IsLicensedVersion` so that we can
reference this property instead of validating the license every time.

Additionally when we're checking for license features, we should be also checking if
the licensed assembly is loaded. When the licensed assembly is not fully
loaded, many features may not work, or are not expected to be used. As
such, this commit updates the logic to check for a valid license as well
as the Assembly to be loaded so that we don't accidentally cause issues
when the assembly is not being used.