(#3224) Update version check tests
Previously the version check would assume that all of the output from
`choco --version` should be considered as part of the version. However,
if the licensed assembly reports it's not valid, it does so very early
in the process, and outputs on this command before the version is
output. In order to work around this, the command to get the running
version should get just the very last line it's returned.