(maint) Extract all used environment variables
These changes extracts all of the environment variables that we
make use of while running Chocolatey CLI. Both variables that are
used as part of running package scripts, and variables we use
only when they are set on the system.

A change to make these environment variables and all other internal
types available to other products and the test projects has been made.