(#1485) Display package source during install/upgrade
This logs the package source uri to info before downloading nupkgs
during both install and upgrade. Additionally, it debug logs the
download URI that the nupkg available at.

The information outputs for both individual package installs/upgrades,
and for dependencies that get installed/upgraded by their parent
package. This information is especially useful when multiple sources
are in play, so the user can determine where individual packages
are coming from.