(#3264) Ignore lock folders in cache directories
As part of the changes that is coming in our custom flavor
of NuGet.Client, the cache directories will now contain the
a new folder where lock files are located when issuing
HTTP requests.

The cache command needs to ignore these directories as
they may contain locked files that can not be cleaned up,
and is not part of the cached HTTP queries.