(#3292) Ensure pre-release flag is used for parent lookup
These changes makes sure that we pass along any pre-release flag
when there is a pre-release package that is a parent of one of the
packages that we are installing, and to verify that the results is not
null before we attempt to use it.

This ensures that pre-release only packages can be found when we
look up the parent, and if by some reason the package can not be
found we do not throw an undesired null exception in this case.