(#3315) Save package contents deployment location
Saves the location of where the contents of the Chocolatey package
(whether that is a native installer, or an extraction of a zip file, or
whatever else is in the package) is deployed to into a
".deploymentLocation" file inside the package info directory.

This allows Chocolatey CLI or GUI to later retrieve it and display it to
the user when listing the details of locally installed package(s).