(#2712) Normalize ExitCode pester tests
We haven't been consistent when checking the exit code of the Chocolatey
command run. This commit updates all instances where we're testing
ExitCode on the output object to also include the String propery if the
ExitCode doesn't match. This allows us to better deduce issues as we
have the Chocolatey output as well.