(#3381) Add new rule command
This adds a new command called ´rule` with an alias of `rules` to be
able to view any Validation Rules that have been implemented by
Chocolatey CLI, or the details of a specific view.

This allows a user to see what is implemented when they are packaging
their Chocolatey CLI package. This is intended to be used in the future
to enhance other functionality.