(maint) Remove unsupported command from tests
The Common Helpers module that we use in the Pester tests includes a
line to extract the list of packages on the system. This commit is
still using the Chocolatey CLI 1.x flags of the choco list command.
This commit removes the no longer supported parameter so that the
output is actually useful.