(maint) Add ability for Agent debugging
Chocolatey Agent adds `--run-actual` on every Chocolatey CLI command
run. When we're in debug mode and `--run-actual` is passed to Chocolatey
CLI, we should not pause execution as the end as we're likely in agent
and cannot acknowledge the pause.

This also adds a change to the cake recipe to add `--run-actual` on the
`unpackself` step as when we're building a debug version we don't want
to pause for the user to press a key.