(#3426) Add deprecation notice for unpackself
Based on internal conversations, while the unpackself command is
currently being used as part of the overall build process, it is not
something that is "required". So, rather than leave this command in
place, let's deprecate it, and then remove the command completely in
the next major version of Chocolatey CLI.

With this command removed, it will be necessary to change the build
process to sign the PowerShell files earlier in the build process than
they currently are, but it shouldn't impact on the end result at all.