(doc) Add missing tab completions for info command
While reviewing the PR to include the deployment location to the choco
info and list commands, it was noticed that there were some tab
completions for the info command missing, specifically the --local-only
option.  On further review, it was clear that there were some
additional options missing, so this commit adds in the missing options
for the info command.

A larger scope of work would be to review all the commands to ensure
that all available options are included, and better yet, it would be
good to generate this tab completion file automatically as part of the
build process.