(maint) Update FindPackage to use base cache class
This updates the method FindPackage to make use of the base class for
the cache context. This is done as there are other calling methods that
now makes use of this function that already specifies this base class,
and to prevent an unnessesary cast this method should be changed as
well.

This is a non-breaking change due to it being the base class of the
existing cache context that was used.