(#2050) Fix broken integration tests
In a previous commit
(https://github.com/chocolatey/choco/commit/fb22630c4bead6f53a8361b66eec64db44d35525)
a change was made to remove the pinned source dependencies from the
list of available packages only when we are not ignoring pinned
packages.  While this change was valid, it forgot to populate the
removedSources property with the results of the call to
RemovePinnedSourceDependencies.  This was an oversight, and although
the change was tested locally, a complete execution of all the
integration tests was not performed, which would have caught this
problem.