(#1485) Fix integration tests on non-Windows
This commit adds the attributes to make the tests only execute on
Windows.  Given the usage of file:// which is a Windows only construct,
these tests will never work on non-Windows.  The check that is being
done is a simple sanity check that the output is being created, so
running this only for Windows is enough.