(build) Stop signing artifacts on some builds
Based on changes to how the Authenticode signing works (namely the need
to use a hardware HSM, or a software as a service based solution), we
need to look to prevent unnecessary signing of files happening.
Signing of files is not required on the scheduled Integration Testing
and SonarQube builds, so we are removing it from these builds.  This is
done using the command line options for:

 --shouldAuthenticodeSignMsis=false
--shouldAuthenticodeSignOutputAssemblies=false
--shouldAuthenticodeSignPowerShellScripts=false