(build) Update to latest Recipe package
Going forward, rather than signing PowerShell scripts on every build,
which is a bit of a waste of effort, we are only going to sign the
scripts when required.  The changes required to to do have been
implemented in 0.28.0 of Chocolatey.Cake.Recipe under issue #149.

This commit updates the Chocolatey CLI build to make use of it.  This
includes changes the DAG for tasks to depend on the new
Verify-PowerShellScripts task, rather than the Sign-PowerShellScripts
task.  In addition, a new Func is created to find the scripts that need
to be verified, since this can, and is, sometimes different from the
list of files that need to be signed.