(build) Add Script Signing build configuration
This new step is the mechanism through which the (non-build related)
PowerShell scripts in this project will be signed from now on. The
result of this build will be any scripts that required signing being
uploaded as artifacts on the build server so that they can be committed
into the repository.

This build configuration will be triggered by changes to the develop
branch only when there has been a change to the target PowerShell scripts.

Additionally, this commit removes the "PrepareSigningEnvironment" meta
runner from the build configurations that are now explicitly no longer
doing any Authenticode signing: ChocolateySchd and ChocolateyQA.