(maint) Made line endings and encoding consistent
UTF-8 with BOM and CFLF.

During the process of bringing in the newly signed PowerShell scripts
from the new build configuration, it was found that the new files had
differnt line endings and encoding compared to the signed files.

Rather than do a commit with included both the applying of the signature
and changes to the source files, decided to do a standalone commit
where the line endings and encodings was made consistent.