(#1761) Update tests based on changes
When we introduced the concept of an enhanced exit code for the choco
feature command, we didn't update the Pester tests that are explicitly
looking for an exit code of 0 when running the choco feature command.

In some circumstances, the useEnhancedExitCodes feature of Chocolatey
CLi is enabled, and the exit code can be 2, rather than 0, so the
assertion had to be updated to use -notin, rather than a straight -ne.

In addition, the output of the Write-Host was changed to made it
clearer that the exit code was actually the exit code, and that the
Chocolatey Output fro the executing command was what it said it was.