(build) Update tests changes to signing
Going forward, the official signing certificate will only be in play on
tagged builds.  As such, we can't easily assert on the certificate
thumbprint. Instead, let's check on the Issuer name, and if the
official one, check for certificate being valid, otherwise let's assert
on it not being correct.  This is similar to what is now done elsewhere
when checking for signatures, i.e. in CCR and Chocolatey.Cake.Recipe.