(test) Output complete string when invalid assert
When trying to understand why an assertion hasn't worked, it is useful
to be able to see all of the output string.  This is accomplished by
using -Because $Output.String, so this has been added around the choco
push tests, to get a better understanding of why the push tests are
intermittently failing.