(maint) Fixed inconsistencies in tab completions
There is an inconsistency in the way that some options are tab
completed. The majority of times we use --name='' as an example, where
the option includes the = sign and the ''. In some places we aren't
doing that.

This commit addresses that issue, by being consistent everywhere.  This
change came about as a result of writing some Pester tests to verify
the output was correct.