(maint) Ensure PowerShell certificate is trusted
During pairing with gep13, discovered that some of the Pester tests
failed in Team City as they are likely waiting for the user to confirm
running the PowerShell scripts as the certificate used to sign choco.exe
differs from the ps1 files on development builds. This updates the tests
to trust the certificate of the ps1 files as well so that the prompt
doesn't happen.