(maint) Update Test Developer Experience
Enhance the local testing process for developers. Instead of building
all nuspec files we find, filter out ones that may have been built as
part of previous tests.

Update the vagrantfile to copy over changed files when running the test
provisioner, but only purge the tests directory.

Update the vagrantfile to set a hostname on the VM to prevent on-screen
error messages that don't seem to affect anything.