(#1761) Add test for new enhanced exit code
Now, when you have useEnhancedExitCodes turned on, if you try to
enable/disable a feature that is already in the desired state, the exit
code will change to be a 2, giving a clear indication that no action was
taken.

This commit adds a Pester test to verify that this works as expected.