(#1760) Add test for new enhanced exit code
Now, when you have useEnhancedExitCodes turned on, if you try to
unset a config value that doesn't exist, or try to set a config value
where the value is already set to that value, the exit code will change
to be a 2, giving a clear indication that no action was taken.

This commit adds a Pester test to verify that this works as expected.